requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
PyMySQL
pydantic
PyJWT
pwdlib[bcrypt]
Pillow
python-multipart
